Factor out quick fault functionality.
I tried a few approaches to factoring a more re-usable version of
quickfault out of the example and settled on this commit. The major
downside is that you have to pass `q` into the property test function
as part of the closure. This seems prone to mistakes, but I don't see
a good way of getting it all into one place.